refactor: move keypair to separate file
